backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
supabase==2.5.1
python-dotenv==1.0.1
pydantic==2.7.4
pydantic-settings==2.3.1
